# cache-require-paths