--- a/package.json
+++ b/package.json
@@ -92,6 +92,7 @@
     "bat-publisher": "^2.0.3",
     "bignumber.js": "^4.0.4",
     "bloodhound-js": "brave/bloodhound",
+    "cache-require-paths": "^0.3.0",
     "clipboard-copy": "^1.0.0",
     "compare-versions": "^3.0.1",
     "deepmerge": "^2.0.1",