@ayumi 2017-12-21
@bsclifton thanks for the reminder! I rebased and tried it out again on Linux (dev and built package)