@bsclifton 2017-12-21
Security review created https://github.com/brave/internal/issues/161

Will wait for results before merging