@diracdeltas 2017-12-21
i wonder if we will run into cache invalidation issues? https://github.com/bahmutov/cache-require-paths/issues/6