@bsclifton 2017-12-21
@ayumi do you know if the cache is invalidated on each run? If not, we could encounter the problem @diracdeltas linked to when running from source (we wouldn't run into it for packaged builds though)