@ayumi 2017-12-21
@diracdeltas @bsclifton That's a good catch. The only cache invalidation happens when the requiring a cached path fails, in that case it resolves from scratch: https://github.com/bahmutov/cache-require-paths/blob/master/index.js#L40

So it sounds like this may affect dev builds, however wouldn't affect packaged builds.
Would it be sufficient to add a post `npm install` hook to delete the cache file?